Re #2407 Move exceptions out of Stack.Types.Build
Also move out ConfigureOpts (and related), IsMutable, ParentMap and UnusedFlags.
